ARO: Move reused functions to the util file
Moving the functions that are used by the file generators to the
util file in the bootstrap package.
